requirements.txt
pytorch-lightning==1.9.0
PyYAML==6.0
tensorboard>=2.2.0
torch>=1.6.1
torchsummary==1.5.1
torchvision>=0.10.1